# websrc/controllers/migrationNCPHandlers.go
/*
Copyright 2023 The Cloud-Barista Authors.

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

	http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
*/
package controllers

import (
	"net/http"
	"time"

	"github.com/cloud-barista/mc-data-manager/models"
	"github.com/labstack/echo/v4"
)

// MigrationNCPToLinuxPostHandler godoc
//
//	@ID 			MigrationNCPToLinuxPostHandler
//	@Summary		Migrate data from NCP to Linux
//	@Description	Migrate data stored in NCP Object Storage to a Linux-based system.
//	@Tags			[Migrate]
//	@Accept			json
//	@Produce		json
//	@Param			RequestBody		body	models.MigrateTask	true	"Parameters required for migration"
//	@Success		200			{object}	models.BasicResponse	"Successfully migrated data"
//	@Failure		400			{object}	models.BasicResponse	"Invalid Request"
//	@Failure		500			{object}	models.BasicResponse	"Internal Server Error"
//	@Router			/migration/ncp/linux [post]
func MigrationNCPToLinuxPostHandler(ctx echo.Context) error {

	start := time.Now()

	logger, logstrings := pageLogInit(ctx, "migncplin", "Export ncp data to linux", start)

	if !osCheck(logger, start, "linux") {
		return ctx.JSON(http.StatusBadRequest, models.BasicResponse{
			Result: logstrings.String(),
			Error:  nil,
		})
	}

	params := models.MigrateTask{}
	if !getDataWithBind(logger, start, ctx, &params) {
		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
			Result: logstrings.String(),
			Error:  nil,
		})
	}

	ncpOSC := getS3COSC(logger, start, "mig", params.SourcePoint)
	if ncpOSC == nil {
		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
			Result: logstrings.String(),
			Error:  nil,
		})
	}

	if !oscExport(logger, start, "ncp", ncpOSC, params.TargetPoint.Path) {
		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
			Result: logstrings.String(),
			Error:  nil,
		})
	}

	jobEnd(logger, "Successfully migrated data from ncp objectstorage to linux", start)
	return ctx.JSON(http.StatusOK, models.BasicResponse{
		Result: logstrings.String(),
		Error:  nil,
	})
}

// MigrationNCPToWindowsPostHandler godoc
//
//	@ID 			MigrationNCPToWindowsPostHandler
//	@Summary		Migrate data from NCP to Windows
//	@Description	Migrate data stored in NCP Object Storage to a Windows-based system.
//	@Tags			[Migrate]
//	@Accept			json
//	@Produce		json
//	@Param			RequestBody		body	models.MigrateTask	true	"Parameters required for migration"
//	@Success		200			{object}	models.BasicResponse	"Successfully migrated data"
//	@Failure		400			{object}	models.BasicResponse	"Invalid Request"
//	@Failure		500			{object}	models.BasicResponse	"Internal Server Error"
//	@Router			/migration/ncp/windows [post]
func MigrationNCPToWindowsPostHandler(ctx echo.Context) error {

	start := time.Now()

	logger, logstrings := pageLogInit(ctx, "migncpwin", "Export ncp data to windows", start)

	if !osCheck(logger, start, "windows") {
		return ctx.JSON(http.StatusBadRequest, models.BasicResponse{
			Result: logstrings.String(),
			Error:  nil,
		})
	}

	params := models.MigrateTask{}
	if !getDataWithBind(logger, start, ctx, &params) {
		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
			Result: logstrings.String(),
			Error:  nil,
		})
	}

	ncpOSC := getS3COSC(logger, start, "mig", params.SourcePoint)
	if ncpOSC == nil {
		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
			Result: logstrings.String(),
			Error:  nil,
		})
	}

	if !oscExport(logger, start, "ncp", ncpOSC, params.TargetPoint.Path) {
		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
			Result: logstrings.String(),
			Error:  nil,
		})
	}

	// migration success. Send result to client
	jobEnd(logger, "Successfully migrated data from ncp to windows", start)
	return ctx.JSON(http.StatusOK, models.BasicResponse{
		Result: logstrings.String(),
		Error:  nil,
	})
}

// MigrationNCPToS3PostHandler godoc
//
//	@ID 			MigrationNCPToS3PostHandler
//	@Summary		Migrate data from NCP to AWS S3
//	@Description	Migrate data stored in NCP Object Storage to AWS S3.
//	@Tags			[Migrate]
//	@Accept			json
//	@Produce		json
//	@Param			RequestBody		body	models.MigrateTask	true	"Parameters required for migration"
//	@Success		200			{object}	models.BasicResponse	"Successfully migrated data"
//	@Failure		500			{object}	models.BasicResponse	"Internal Server Error"
//	@Router			/migration/ncp/aws [post]
func MigrationNCPToS3PostHandler(ctx echo.Context) error {

	start := time.Now()

	logger, logstrings := pageLogInit(ctx, "migncps3", "Export ncp data to s3", start)

	params := models.MigrateTask{}
	if !getDataWithBind(logger, start, ctx, &params) {
		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
			Result: logstrings.String(),
			Error:  nil,
		})
	}

	ncpOSC := getS3COSC(logger, start, "mig", params.SourcePoint)
	if ncpOSC == nil {
		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
			Result: logstrings.String(),
			Error:  nil,
		})
	}

	awsOSC := getS3OSC(logger, start, "mig", params.TargetPoint)
	if awsOSC == nil {
		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
			Result: logstrings.String(),
			Error:  nil,
		})
	}

	logger.Info().Msgf("Start migration of NCP Object Storage to AWS S3")
	if err := ncpOSC.Copy(awsOSC); err != nil {
		end := time.Now()
		logger.Error().Msgf("OSController migration failed : %v", err)
		logger.Info().Msgf("End time : %s", end.Format("2006-01-02T15:04:05-07:00"))
		logger.Info().Msgf("Elapsed time : %s", end.Sub(start).String())
		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
			Result: logstrings.String(),
			Error:  nil,
		})
	}

	// migration success. Send result to client
	jobEnd(logger, "Successfully migrated data from ncp to s3", start)
	return ctx.JSON(http.StatusOK, models.BasicResponse{
		Result: logstrings.String(),
		Error:  nil,
	})
}

// MigrationNCPToGCPPostHandler godoc
//
//	@ID 			MigrationNCPToGCPPostHandler
//	@Summary		Migrate data from NCP to GCP Cloud Storage
//	@Description	Migrate data stored in NCP Object Storage to GCP Cloud Storage.
//	@Tags			[Migrate]
//	@Accept			json
//	@Produce		json
//	@Param			RequestBody		body	models.MigrateTask	true	"Parameters required for migration"
//	@Param 			gcpCredential	formData 	file 			false 	"Parameters required to generate test data"
//	@Success		200			{object}	models.BasicResponse	"Successfully migrated data"
//	@Failure		500			{object}	models.BasicResponse	"Internal Server Error"
//	@Router			/migration/ncp/gcp [post]
func MigrationNCPToGCPPostHandler(ctx echo.Context) error {

	start := time.Now()

	logger, logstrings := pageLogInit(ctx, "migncpgcp", "Export ncp data to gcp", start)

	params := models.MigrateTask{}
	if !getDataWithBind(logger, start, ctx, &params) {
		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
			Result: logstrings.String(),
			Error:  nil,
		})
	}

	ncpOSC := getS3COSC(logger, start, "mig", params.SourcePoint)
	if ncpOSC == nil {
		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
			Result: logstrings.String(),
			Error:  nil,
		})
	}

	gcpOSC := getGCPCOSC(logger, start, "mig", params.TargetPoint)
	if gcpOSC == nil {
		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
			Result: logstrings.String(),
			Error:  nil,
		})
	}

	logger.Info().Msgf("Start migration of NCP Object Storage to GCP Cloud Storage")
	if err := ncpOSC.Copy(gcpOSC); err != nil {
		end := time.Now()
		logger.Error().Msgf("OSController migration failed : %v", err)
		logger.Info().Msgf("End time : %s", end.Format("2006-01-02T15:04:05-07:00"))
		logger.Info().Msgf("Elapsed time : %s", end.Sub(start).String())
		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
			Result: logstrings.String(),
			Error:  nil,
		})
	}

	jobEnd(logger, "Successfully migrated data from ncp to gcp", start)
	return ctx.JSON(http.StatusOK, models.BasicResponse{
		Result: logstrings.String(),
		Error:  nil,
	})
}